+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+### Added
+- You can now specify the size of researches in the research tree
+
 ### Changed
 - Re-enabled REI integration
+- Changed default amount in fireworks rewards to 1 (from 0) 
+
+### Fixed
+- Fixed broken JEI mixin